Refactor: 미사용 import삭제 및 오타 수정 (#14)
