+++ src/api/endpoints/VerifyCode.ts
@@ -2,7 +2,7 @@
 import { apiClient } from '../Client';
 import { BaseResponseDto } from '../dto/BaseResponse.dto';
 
-export const postVeifyCode = async (
+export const postVerifyCode = async (
   verifyCode: string,
 ): Promise<BaseResponseDto<void>> => {
   const res = await apiClient<BaseResponseDto<void>>(
